pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["korotkov_egor <korotkovegor10@gmail.com>"]
readme = "README.md"
packages = [
  { include = "difference_calculator" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "difference_calculator.scripts.gendiff:show_certificate"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
